feat: modularize safety UI dialogs
