+++ gui/dialogs/__init__.py
@@ -10,9 +10,12 @@ _module: ModuleType | None = None
 
 def __getattr__(name: str):
     global _module
+    if name == "FMEARowDialog":
+        from .fmea_row_dialog import FMEARowDialog as dlg
+        return dlg
     if _module is None:
         _module = importlib.import_module("mainappsrc.automl_core")
     return getattr(_module, name)
 
 
-__all__: list[str] = []
+__all__ = ["FMEARowDialog"]
